[refactor] 게임 로직과 무관한 메소드 분리
게임 로직과 관련 없는 메소드를 Utility.cs 파일로 분리하여 관리
